// Assets-Uncapped/front-portal-develop/src/services/api/collections/models/GoCardlessMandateResponse.ts
export interface GoCardlessMandateResponse {
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    status?: GoCardlessMandateResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    iban?: string;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    bank?: string;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    country?: string;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    currency?: string;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    bankAccountNumberTail?: string;
    /**
     * 
     * @type {string}
     * @memberof GoCardlessMandateResponse
     */
    gcInstance?: GoCardlessMandateResponseGcInstanceEnum;
}


/**
 * @export
 */
export const GoCardlessMandateResponseStatusEnum = {
    PendingCustomerApproval: 'PENDING_CUSTOMER_APPROVAL',
    PendingSubmission: 'PENDING_SUBMISSION',
    Submitted: 'SUBMITTED',
    Active: 'ACTIVE',
    Failed: 'FAILED',
    Cancelled: 'CANCELLED',
    Expired: 'EXPIRED',
    Consumed: 'CONSUMED',
    Blocked: 'BLOCKED',
    SuspendedByPayer: 'SUSPENDED_BY_PAYER',
    Unknown: 'UNKNOWN'
} as const;
export type GoCardlessMandateResponseStatusEnum = typeof GoCardlessMandateResponseStatusEnum[keyof typeof GoCardlessMandateResponseStatusEnum];

/**
 * @export
 */
export const GoCardlessMandateResponseGcInstanceEnum = {
    Eur: 'EUR',
    Us: 'US',
    Canada: 'CANADA',
    Marcus: 'MARCUS'
} as const;
export type GoCardlessMandateResponseGcInstanceEnum = typeof GoCardlessMandateResponseGcInstanceEnum[keyof typeof GoCardlessMandateResponseGcInstanceEnum];